Add missing migration condition
AB#7621
AB#7622

+++ src/NHSD.BuyingCatalogue.Database/PreDeployment/MigrateSolutionDetailToCatalogueItem.sql
@@ -171,5 +171,6 @@ GO
     Delete existing solution data
 ------------------------------------------------------------------------*/
 
-DELETE FROM dbo.Solution;
+IF UPPER('$(MIGRATE_TO_CATALOGUE_ITEM)') = 'TRUE'
+    DELETE FROM dbo.Solution;
 GO
